Fixes several minor issues, reorganizes some data, and adds realm support.
There we a couple minor syntax/naming issues from the switch back to standard
python coding conventions which were fixed.

The replay class had its players attribute restructured to have two access
points: replay.players which is an unsorted list of players, and
replay.player which is a dictionary mapping players to their pids.

Basic replay.initData parasing was implemented to gain support for realms.
This addition made it possible to fix the player urls to work across realms!

A small shell.py script was added for debugging purposes, to use it invoke
the interactive prompt with `python -i shell.py` which will execute the setup
code and bring you to the python shell for further work.