Readers now return the data and are documented.
Instead of building the replay structure, readers now return the data to the
caller. This is a more practical and traditional approach. Now readers don't
currently use the replay that gets passed in.  Perhaps they will still need
the replay options in the future but the call signature might need changing.

Addtionally, the commenting is mostly complete for the readers file as well.
Now that __init__.py and readers.py are more or less complete (aside from the
ever pressent TODO's) work will focus on finishing up and documenting the
processors layer of sc2reader.