Fix several bugs related to the initData file.
There appears to be some element of variability in the distance between
the sc_account_id and the map data. Additionally, there are some instances
where the map_data isn't even present in the file at all!

Adjust the reader as well as the processor to account for these new findings.