Apply experimental player load scheme.
Superficial testing says this works better than the previous method
but I still need to make sure.

Also, since 24247 is it apparently possible for computer players to
not have attributes assigned to them. I have no idea how we are
supposed to pick the team they are on without the attributes.