@mischanix wrote this back in Sept 2015 but I neglected to commit until now.
He wrote: "I decided to spring for 64-bit support, so lots of stuff with the dump
script changed.  Lots of ID changes as well since HotS, it seems.
Shame they still haven't shipped an editor capable of dumping the LotV
balance data."

However in Nov 2015, @icza writes: "LotV is released, and balance data export is available for LotV. Closing the issue"

In any case, committing this in case it's useful in the future.